(refactor) Moved all styles from global component style sheet into localized component style sheets
