+++ client/components/Navigation/LinkRow/style.js
@@ -1,0 +1,19 @@
+module.exports.navigation_body = {
+  display: 'flex',
+  flexDirection: 'row',
+  paddingRight: '50px'
+};
+
+module.exports.navigation_link = {
+  cursor: 'pointer',
+  padding: '15px 30px 15px',
+  fontSize: '24px',
+  color: 'rgba(41, 67, 78, 1.0)',
+  WebkitTransition: '0.2s'
+};
+
+module.exports.navigation_link_hovered = {
+  padding: '9px 30px 15px',
+  color: 'rgba(0, 122, 193, 1.0)',
+  WebkitTransition: '0.2s'
+};
