+++ client/components/Navigation/Title/style.js
@@ -1,0 +1,11 @@
+module.exports.title_left = {
+  cursor: 'pointer',
+  marginLeft: '50px',
+  padding: '0 6px',
+  fontSize: '30px',
+  color: 'rgba(41, 67, 78, 1.0)'
+};
+
+module.exports.title_right = {
+  color: 'rgba(0, 122, 193, 1.0)'
+};
