+++ client/components/Navigation/style.js
@@ -1,0 +1,11 @@
+module.exports.main = {
+  position: 'fixed',
+  width: '100%',
+  height: '70px',
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+  WebkitTransition: 'linear 0.1s',
+  zIndex: '1'
+};
